js: Rename muting_ui module to muted_topics_ui.
