+++ frontend_tests/node_tests/dispatch.js
@@ -32,7 +32,7 @@ const message_edit = mock_esm("../../static/js/message_edit");
 const message_events = mock_esm("../../static/js/message_events");
 const message_list = mock_esm("../../static/js/message_list");
 const message_lists = mock_esm("../../static/js/message_lists");
-const muting_ui = mock_esm("../../static/js/muting_ui");
+const muted_topics_ui = mock_esm("../../static/js/muted_topics_ui");
 const muted_users_ui = mock_esm("../../static/js/muted_users_ui");
 const night_mode = mock_esm("../../static/js/night_mode");
 const notifications = mock_esm("../../static/js/notifications");
@@ -235,7 +235,7 @@ run_test("muted_topics", ({override}) => {
     const event = event_fixtures.muted_topics;
 
     const stub = make_stub();
-    override(muting_ui, "handle_topic_updates", stub.f);
+    override(muted_topics_ui, "handle_topic_updates", stub.f);
     dispatch(event);
     assert.equal(stub.num_calls, 1);
     const args = stub.get_args("muted_topics");
