+++ static/js/hotkey.js
@@ -24,7 +24,7 @@ import * as message_edit from "./message_edit";
 import * as message_flags from "./message_flags";
 import * as message_lists from "./message_lists";
 import * as message_view_header from "./message_view_header";
-import * as muting_ui from "./muting_ui";
+import * as muted_topics_ui from "./muted_topics_ui";
 import * as narrow from "./narrow";
 import * as navigate from "./navigate";
 import * as overlays from "./overlays";
@@ -883,7 +883,7 @@ export function process_hotkey(e, hotkey) {
             return true;
         }
         case "toggle_topic_mute":
-            muting_ui.toggle_topic_mute(msg);
+            muted_topics_ui.toggle_topic_mute(msg);
             return true;
         case "toggle_message_collapse":
             condense.toggle_collapse(msg);
